Backend/requirements.txt
pandas~=2.2.1
fastapi~=0.103.0
sqlalchemy~=2.0.25
openai~=1.9.0
python-dotenv~=0.21.0
pydantic~=2.5.3
alembic~=1.8.1
uvicorn
psycopg2